feat: allow tracking kafka produce and consume offsets (#18)
This was the one remaining feature from the golang library that I
haven't ported yet. Now it's done. It allows tracking Kafka offsets.
Datadog uses this to detect lag and consumers that are offline and not
reporting.

After every publish, we push the last offset to Datadog. After every
consume, we push the last read offset to Datadog. It can then calculate
the offset and alert us on issues.

---------

Co-authored-by: Matt Sutkowski <msutkowski@gmail.com>